change upload to not extract the file data, but to just create a subfolder and store the file there.
the reason for this is because this is just the poc for a file upload
api. i want this to be more of a "self living micro service". Once I
might think this would become something more usable, i may try to figure
out some bespoke changes required for my works use case (i.e. blob
storage or other cloud fs stores) to store the files.